Products now add to a database, viewed on shop.index, used php artisan make:auth to create login and register functions

+++ database/migrations/2016_09_13_224615_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->text('description');
             $table->string('image', 100);
             $table->decimal('price', 5, 2);
-            $table->float('quantity');
+            $table->smallInteger('quantity');
             $table->timestamps();
         });
     }
